Fix find-symbol command after last commit ...
Passing None means "no search string" and
custom_match=is_some_kind_of_lib
together means that the function will return every file from the
environment that is a shared or static library.

Because now we do "if '/' in search_string:...", search_string cannot
be None.

This is inelegant and I should definitely make the find_in_env function
be able to accept a string or None since passing an empty string to
solve the problem is somewhat shameful